Sleep for 30 seconds after a ws connect failure
It's useful to use something like pm2 to auto restart this program after a failure. If you get a single 429 error from trying to connect to the WS api too frequently, this program will error out, and pm2 will immediately restart it. This leads to a never-ending cycle of 429 errors, making it hard to find what originally may have caused the program to connect too frequently.

This commit adds a simple 30 second delay before quitting the program after a WS connect failure. This should give enough time to cool down the rate limiter so the log output contains useful errors, not 1000+ 429 errors.